Rename the PC layer file
